% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbc.counter-package.R
\docType{package}
\name{wbc.counter-package}
\alias{wbc.counter}
\alias{wbc.counter-package}
\title{wbc.counter: Wild Boar Carcasses counter app}
\description{
Application of EFSA Exit Strategy on how to provide cumulative evidence of the absence of African swine fever virus circulation in wild boar populations using standard surveillance measures.
}
\author{
\strong{Maintainer}: Ileana Baldi \email{ileana.baldi@ubep.unipd.it} (\href{https://orcid.org/0000-0002-8578-9164}{ORCID})

Authors:
\itemize{
  \item Federica Loi \email{federica.loi@izs-sardegna.it} (\href{https://orcid.org/Y0000-0003-3045-3150}{ORCID})
}

Other contributors:
\itemize{
  \item Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [contributor]
}

}
\keyword{internal}
